test(fixtures): file the composed programs under the corpus packages
The authored program corpus is split into tests/fixtures/logical for programs
that carry no placement and tests/fixtures/placed for those with an explicitly
placed execution path. The composed roots followed neither: one top-level
module held both kinds.

SameModule / CrossModule / Fused and REFERENCE_PROGRAMS move to
tests/fixtures/logical/hir_composition.py; MoEMegaKernel, whose experts run on
disjoint slices of one CTA topology, moves to
tests/fixtures/placed/moe_mega_kernel.py. Their one importer follows.
